# docs/a7_openapi.yml
openapi: 3.0.0
info:
  version: 1.0.0
  title: 'MelodyAuctions Web API'
  description: 'Web Resources Specification (A7) for MelodyAuctions'
servers:
- url: http://lbaw.fe.up.pt
  description: Production server
...
paths:
...
  /auction/create:
    get:
      operationId: R201
      summary: 'R201: Auction Creation Form'
      description: 'Provide new auction creation form. Access: USR'
      tags:
        - 'M02: Auctions'
      responses:
        '200':
          description: 'Ok. Show auction creation UI'
        '302':
          description: 'Redirect because is an unauthorized user '
          headers:
            Location:
              schema:
                type: string
              examples:
                302Failure:
                  description: 'Unauthorized. You need to be an admin'
                  value: '/login'
...
  /auction/{id}:
    get:
      operationId: R203
      summary: 'R203: View auction'
      description: 'Show the individual auction. Access: PUB'
      tags:
        - 'M02: Auctions'
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: Numeric ID of the auction to get
            
      responses:
        '200':
          description: 'Ok. Show view Auction UI'
        '404':
          description: 'Auction Not found'

    put:
      operationId: R204
      summary: 'R204: Auction Update'
      description: 'Processes the auction update form . Access: OWN ADM'
      tags:
        - 'M02: Auctions'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                name:
                  type: string
                startDate:
                  type: string
                startPrice:
                  type: number
                minBidDif:
                  type: number
                photo:
                  type: string
                  format: binary
                description:
                  type: string
                categoryId:
                  type: integer
                manufactorId:
                  type: integer

      responses:
        '302':
          description: 'Redirect after processing the new auction.'
          headers:
            Location:
              schema:
                type: string
              examples:
                302Success:
                  description: 'Successful update. Redirect to auction UI.'
                  value: '/auction/{id}'
                302Failure:
                  description: 'Failed update. Redirect to auction UI.'
                  value: '/auction/{id}'

    delete:
      operationId: R205
      summary: 'R205: Delete auction'
      description: 'Delete auction. Access: OWN ADM'
      tags:
        - 'M02: Auctions'
      responses:
        '302':
          description: 'Redirect after deleting auction.'
          headers:
            Location:
              schema:
                type: string
              examples:
                302Success:
                  description: 'Successful deletion. Redirect to auctions page'
                  value: '/auction'
                302Failure:
                  description: 'Failed to delete auction. Redirect to auction that wasnt deleted.'
                  value: '/auction/{id}'
        '401':
          description: 'You must be the auction owner or an admin'
  /auction:
    get:
      operationId: R206
      summary: 'R206: View auctions'
      description: 'Show auctions. Access: PUB'
      tags:
        - 'M02: Auctions'
      parameters:
        - in: query
          name: ongoing
          description: boolean with ongoing flag value
          schema:
            type: boolean
          required: false
        - in: query
          name: query
          description: String to use for full-text search
          schema:
            type: string
          required: false
        - in: query
          name: categoryId
          description: id of the category of the auction
          schema:
            type: integer
          required: false
        - in: query
          name: minPrice
          description: minimum bid price
          schema:
            type: integer
          required: false
        - in: query
          name: maxPrice
          description: maximum bid price
          schema:
            type: integer
          required: false
        - in: query
          name: order
          description: magnitude of prices to be displayed
          schema:
            type: integer
          required: false
      responses:
        '200':
          description: 'Ok. Show view Auction UI'
    post:
      operationId: R207
      summary: 'R207: Auction Creation'
      description: 'Processes the new auction creation form submission. Access: OWN'
      tags:
        - 'M02: Auctions'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                name:
                  type: string
                startDate:
                  type: string
                startPrice:
                  type: number
                minBidDif:
                  type: number
                photo:
                  type: string
                  format: binary
                description:
                  type: string
                categoryId:
                  type: integer
                manufactorId:
                  type: integer
              required:
                - name
                - startDate
                - startPrice
                - minBidDif
                - description
                - categoryId
                - manufactorId

      responses:
        '302':
          description: 'Redirect after processing the new auction.'
          headers:
            Location:
              schema:
                type: string
              examples:
                302Success:
                  description: 'Successful addition. Redirect to auction UI.'
                  value: '/auction/{id}'
                302Failure:
                  description: 'Failed authentication. Redirect to auction creation form.'
                  value: '/auction/create'
...