# webapp/app/Http/Controllers/AuctionController.php
<?php

namespace App\Http\Controllers;


use App\Models\Auction;
use App\Models\Category;
use App\Models\Manufactor;


use Illuminate\Support\Str;
use Illuminate\Support\Facades\Auth;
use Illuminate\Support\Facades\DB;
use Illuminate\Http\Request;

use PDOException;
use Carbon\Carbon;	

class AuctionController extends Controller
{
    private function search_auctions_helper(Request $request){
        try{
            $search = $request->search;
            $categoryId = $request->categoryId;
    
            $useSearch = !is_null($search);
            $useCategory = $categoryId !=='-1' && !is_null($categoryId);
            $useOnGoing = $request->ongoing!=='-1' && !is_null($request->ongoing);
            $onGoing = $request->ongoing==1;
            $offset = (int)$request->offset??0;
            DB::beginTransaction();
            DB::statement('SET TRANSACTION ISOLATION LEVEL REPEATABLE READ READ ONLY');

            $auctionsAfterFilter = Auction::when($useSearch,
                                    function($query1) use($search,$categoryId,$onGoing,$useCategory,$useOnGoing){
                                        $query1->when($useCategory,
                                        function($query2) use ($search,$categoryId,$onGoing,$useOnGoing){
                                            $query2->when($useOnGoing,
                                            function($query3) use($search,$categoryId,$onGoing){
                                                    $query3->when($onGoing,
                                                    function($query) use($search,$categoryId){
                                                        $query->whereRaw('tsvectors @@ plainto_tsquery(\'english\', ?) AND category_id=? AND enddate>CURRENT_TIMESTAMP',
                                                        [$search,$categoryId]);
                                                    },
                                                    function($query) use($search,$categoryId){
                                                        $query->whereRaw('tsvectors @@ plainto_tsquery(\'english\', ?) AND category_id=? AND enddate<CURRENT_TIMESTAMP',
                                                        [$search,$categoryId]);
                                                    });
                                            },
                                            function($query) use($search,$categoryId){
                                                $query->whereRaw('tsvectors @@ plainto_tsquery(\'english\', ?) AND category_id=?',
                                                        [$search,$categoryId]);
                                            });
                                        },
                                        function($query2) use($search,$onGoing,$useOnGoing){
                                            $query2->when($useOnGoing,
                                            function($query3) use($search,$onGoing){
                                                    $query3->when($onGoing,
                                                    function($query) use($search){
                                                        $query->whereRaw('tsvectors @@ plainto_tsquery(\'english\', ?) AND enddate>CURRENT_TIMESTAMP',
                                                        [$search]);
                                                    },
                                                    function($query) use($search){
                                                        $query->whereRaw('tsvectors @@ plainto_tsquery(\'english\', ?) AND enddate<CURRENT_TIMESTAMP',
                                                        [$search]);
                                                    });
                                            },
                                            function($query) use($search){
                                                $query->whereRaw('tsvectors @@ plainto_tsquery(\'english\', ?)',
                                                        [$search]);
                                            });
                                        });
                                    },
                                    function($query1) use($categoryId,$onGoing,$useCategory,$useOnGoing){
                                        $query1->when($useCategory,
                                        function($query2) use($categoryId,$onGoing,$useOnGoing){
                                            $query2->when($useOnGoing,
                                            function($query3) use($categoryId,$onGoing){
                                                    $query3->when($onGoing,
                                                    function($query) use($categoryId){
                                                        $query->whereRaw('category_id=? AND enddate>CURRENT_TIMESTAMP',
                                                        [$categoryId]);
                                                    },
                                                    function($query) use($categoryId){
                                                        $query->whereRaw('category_id=? AND enddate<CURRENT_TIMESTAMP',
                                                        [$categoryId]);
                                                    });
                                            },
                                            function($query) use($categoryId){
                                                $query->whereRaw('category_id=?',
                                                [$categoryId]);
                                            });
                                        },
                                        function($query2) use($onGoing,$useOnGoing){
                                            $query2->when($useOnGoing,
                                            function($query3) use($onGoing){
                                                    $query3->when($onGoing,
                                                    function($query){
                                                        $query->whereRaw('enddate>CURRENT_TIMESTAMP');
                                                    },
                                                    function($query){
                                                        $query->whereRaw('enddate<CURRENT_TIMESTAMP');
                                                    });
                                            });
                                        });
                                    }

                                );

            $nrAuctions = $auctionsAfterFilter->count();
            
            $auctions = $auctionsAfterFilter->selectRaw('id,
                                                        enddate,
                                                        CASE 
                                                        WHEN CURRENT_TIMESTAMP < enddate
                                                            THEN 1
                                                            ELSE 0
                                                            END
                                                        AS active,
                                                        name AS productName,
                                                        currentprice + minbidsdif AS minBid,
                                                        photo')
                                            ->when($useSearch,function($query,$search){
                                                $query->orderByRaw('ts_rank(tsvectors, plainto_tsquery(\'english\', ?)) DESC,name', [$search]);
                                            },function($query){
                                                $query->orderBy('name');
                                            }
                                            )
                                            ->offset($offset*10)
                                            ->limit(10)
                                            ->get();
            DB::commit();
        }
        catch(PDOException $e){
            DB::rollBack();
        }
        return [$auctions,$nrAuctions];
    }
    public function search_results_html(Request $request){
        [$auctions,$nrAuctions]=$this->search_auctions_helper($request);
        $offset = ((int)$request->offset) ?? 0;
        

        return  view('partials.auctions')
                ->with('auctions',$auctions)
                ->with('nrAuctions',$nrAuctions)
                ->with('offset',$offset);
               
                
    }

    public function search_results_json(Request $request){
        $query = $request->search;
        return  response()->json(Auction::select('name','description','id')
                                        ->whereRaw('tsvectors @@ plainto_tsquery(\'english\', ?)', [$query])
                                        ->orderByRaw('ts_rank(tsvectors, plainto_tsquery(\'english\', ?)) DESC', [$query])
                                        ->take(10)
                                        ->get());
    }

    public function list(Request $request){
        $search = $request->search;
        $offset = ((int)$request->offset) ?? 0;
        [$auctions,$nrAuctions]=$this->search_auctions_helper($request);
        
        $categories = Category::select('name','id')
                                ->orderBy('name')
                                ->get();

        $categoryName = Category::select('name')
                                ->where('id','=',$request->categoryId)
                                ->get();
        
                                
        if(sizeof($categoryName)<=0)
            $categoryName = "Any category";
        else{
            $categoryName = $categoryName[0]->name;
        }

        $onGoingText = $request->ongoing==='1'?'Active':($request->ongoing==='0'?'Closed':'None selected');
        return view('pages.auctions')
            ->with('nrAuctions',$nrAuctions)
            ->with('auctions',$auctions)
            ->with('search',$search)
            ->with('onGoing',$request->ongoing)
            ->with('onGoingText',$onGoingText)
            ->with('categoryId',$request->categoryId)
            ->with('categoryName',$categoryName)
            ->with('categories',$categories)
            ->with('offset',$offset);
    }

    /**
     * Show the form for editing the specified resource.
     *
     * @param  \App\Models\Auction  $auction
     * @return \Illuminate\Http\Response
     */
    public function edit($auctionId)
    {
        if (Auth::check()){
            $user_id = Auth::id();
            $auction = Auction::find($auctionId);
            if ($auction->owner_id == $user_id){
                return view('pages.auction_edit', ['auction' => $auction]);
            }
        }
        return redirect()->route('home');
    }


    /**
     * Update the specified resource in storage.
     *
     * @param  \Illuminate\Http\Request  $request
     * @param  \App\Models\Auction  $auction
     * @return \Illuminate\Http\Response
     */

    public function ownerUpdate(Request $request, $auctionId)
    {
        if (Auth::check()){
            $user_id = Auth::id();
            $auction = Auction::find($auctionId);
            if ($auction->owner_id == $user_id){
                $success = true;
                $auction = Auction::find($auctionId);

                $validated = $request->validate([
                    'name' => 'required|max:50',
                    'description' => 'required',
                    'minBidDif' => 'required|numeric|min:0.01',
                    'startDate' => 'required|date',
                    'endDate' => 'required|date',
                ]);

                $auction->name = $request->input('name');
                $auction->description = $request->input('description');
                $auction->minbidsdif = $request->input('minBidDif');

                // Start date
                $inputStartDate = Carbon::parse($request->input('startDate'))->format('Y-m-d H:i:s.u'); // get the start date from the form
                $inputStartDate = substr($inputStartDate, 0, -3); // remove last 3 digits to be according to the dates saved in the database
                $nowDate = Carbon::now()->format('Y-m-d H:i:s.u'); // get current date
                $nowDate = substr($nowDate, 0, -3); // remove last 3 digits to be according to the dates saved in the database
            
                // check if the input date and the auction start date is in the future //check if input date is the current one stored in the database
                if ((Carbon::parse($inputStartDate)->gt(Carbon::now()) && Carbon::parse($auction->startdate)->gt(Carbon::now())) || $auction->startdate == $inputStartDate) { 
                    $auction->startdate = $inputStartDate; 
                } else {
                    $success = false;
                }

                // End date
                $inputEndDate = Carbon::parse($request->input('endDate'))->format('Y-m-d H:i:s.u'); // get the end date from the form
                $inputEndDate = substr($inputEndDate, 0, -3); // remove last 3 digits to be according to the dates saved in the database

                // check if the input date and the auction start date is in the future //check if input date is the current one stored in the database
                if ((Carbon::parse($inputEndDate)->gt(Carbon::now()) && Carbon::parse($auction->startdate)->gt(Carbon::now())) || $auction->enddate == $inputEndDate) { 
                    $auction->enddate = $inputEndDate; 
                } else {
                    $success = false;
                }

                
                if ($success){
                    $auction->save();
                }
                return view('pages.auction_edit', ['auction' => $auction, 'success' => $success]);
            }
        }
        return redirect()->route('home');
    }
    public function ownerDelete($auctionId)
    {
        if (Auth::check()){
            $user_id = Auth::id();
            $auction = Auction::find($auctionId);
            if ($auction->owner_id == $user_id){
                $auction->delete();
                return redirect('/');
            }
        }
        return redirect()->route('home');
    }

    public function updatePhoto(Request $request,$auctionId)
    {
        if (Auth::check()){
            $user_id = Auth::id();
            $auction = Auction::find($auctionId);
            if ($auction->owner_id == $user_id){
                $validated = $request->validate([
                    'photo' => 'required|image|mimes:jpeg,png,jpg,gif,svg|max:2048',
                ]);
                $image_path = $request->file('photo')->store('auctions','public');
                $auction->photo = $image_path;
                $auction->save();
                return redirect()->route('auction.edit',$auctionId);
            }
        }
        return redirect()->route('home');
    }



    public static function update(Request $request, $id)
    {
        $auction = Auction::find($id);
        $auction->name = $request->input('name');
        $auction->description = $request->input('description');
        $auction->save();
        return redirect('admin/auctions');
    }

    public static function destroy($id)
    {
        $auction = Auction::find($id);
        $auction->delete();
        return redirect('admin/auctions');
    }
  
    public static function find($id)
    {
        $auction = Auction::find($id);
        return $auction;
    }

    public static function findAll()
    {
        $auctions = Auction::all();
        return $auctions;
    }
    
    private function uploadImage($request ){
        //TODO resize img
       
        return $request->file('photo')->store('image', 'public');
    }

    /**
     * Show the form for creating a new auction.
     *
     * @return \Illuminate\Http\Response
     */
    public function create(Request $request)
    {
        //$this->authorize('create', Auction::class);
        return view('pages.create_auction', ['categories' => Category::all(), 'manufactors' => Manufactor::all()]);
    }

    /**
     * Store a newly created resource in storage.
     *
     * @param  \Illuminate\Http\Request  $request
     * @return \Illuminate\Http\Response
     */
    public function store(Request $request)
    {
        $this->validate($request, [            
            'photo' => 'required|image|mimes:jpg,png,jpeg',
            'name' => 'required',
            'description' => 'required',
            'manufactor' => 'required',
            'category' => 'required|integer|min:0',
            'startdate' => 'required|date',
            'enddate' => 'required|date',
            'startvalue' => 'required|numeric',
            'minbiddiff' => 'required|numeric',
        ]);

        $image_path = "";
        try{
            $errors = [];

            error_log("sss");
            $auction = new Auction();

            $auction->name = $request->input('name');
            $auction->description = $request->input('description');
            $auction->startprice = floatval($request->input('startvalue'));
            $auction->minbidsdif = floatval($request->input('minbiddiff'));


            $auction->startdate = $request->input('startdate');
            $auction->enddate = $request->input('enddate');

            $auction->owner_id = 2; //TODO: mudar para o id do user logado

            //$this->authorize('store', $auction);

            $now = date('Y-m-d\TH:i:s');

            
            DB::beginTransaction();
            DB::statement('SET TRANSACTION ISOLATION LEVEL SERIALIZABLE');

            if($auction->startdate < $now){
                $errors['startdate_error'] = 'Start date must be after now!';
                throw new PDOException('startdate error!');
            }

            if($auction->enddate < $auction->startdate){
                $errors['enddate_error'] = 'End date must be after Star Date!';
                throw new PDOException('enddate error!');
            }

            if($auction->startprice < 0.01){
                $errors['start_value'] = 'Start value must be at least 0.01!';
                throw new PDOException('start_value error!');
            }
                
            if($auction->minbidsdif < 0.01){
                $errors['minbiddiff'] = 'Min Bid difference must be at least 0.01!';
                throw new PDOException('minbiddiff error!');
            }

            

            if(!Category::find(intval($request->input('category')))->exists()){
                $errors['category_error'] = 'Category does not exists!';
                throw new PDOException('category error!');
            }

            $auction->category_id = intval($request->input('category'));

            $manufactor_name = ucfirst(strtolower($request->input('manufactor')));
            $manufactor = Manufactor::where('name', $manufactor_name)->first();
            if($manufactor === null){
                $manufactor = Manufactor::firstOrCreate(['name' => $manufactor_name, 'id' => Manufactor::max('id')+1]);
            }

            $auction->manufactor_id = $manufactor->id;
            
            $auction->photo = AuctionController::uploadImage($request);
            $auction->id = Auction::max('id') + 1;


            $auction->save();
            
            DB::commit();
        }catch(PDOException $e){

            error_log($e->getMessage());
            if($auction->photo!==""){
                // apagar imagem  
            }

            DB::rollBack();
            return redirect('auction/create')->withErrors($errors);
        }

        return redirect('auction/'.$auction->id);


    }

    /**
     * Display the auction page.
     *
     * @param  \App\Models\Auction  $auction
     * @return \Illuminate\Http\Response
     */
    public function show(int $id)
    {
        $auction = Auction::find($id);
        if($auction){
            //$this->authorize('show', $auction);
            return view('pages.auction', ['auction' => $auction]);
        }else{
            abort(404);
        }
    }

    public function bids(Request $request, $id){

        $offset =$request->offset;
        if($offset!==null){
            // Validate if a string is a valid number

            if(! preg_match('/^[0-9]+$/', $offset)){
                abort(404);
            }   
        }

        $auction = Auction::find($id);
        if($auction==null){
            abort(404);
        }

        $bids = [];
        if($offset!==null){
            $bids = $auction->bids_offset(intval($offset));
        }else{
            $bids =$auction->bids;
        }
        
        error_log($bids);
        return view('partials.bids', ['bids' => $bids]);
    }

}